ICU-1007 Need to do 8 byte alignment for the 64-bit Solaris compiler
X-SVN-Rev: 5560

+++ icu4c/source/tools/gennorm/store.c
@@ -65,7 +65,7 @@ static uint16_t indexes[_NORM_INDEX_TOP]={ 0 };
 typedef struct UToolMemory {
     char name[64];
     uint32_t count, size, index;
-    uint32_t array[1];
+    double array[1];
 } UToolMemory;
 
 static UToolMemory *
